feat: add helper functions to Account object
